# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Memoh API",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Validate user credentials and issue a JWT",
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/bots/{id}/channel/{platform}": {
            "get": {
                "description": "Get bot channel configuration",
                "tags": [
                    "bots"
                ],
                "summary": "Get bot channel config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bot ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Channel platform",
                        "name": "platform",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/channel.ChannelConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update bot channel configuration",
                "tags": [
                    "bots"
                ],
                "summary": "Update bot channel config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bot ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Channel platform",
                        "name": "platform",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Channel config payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/channel.UpsertConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/channel.ChannelConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/channels": {
            "get": {
                "description": "List channel meta information including capabilities and schemas",
                "tags": [
                    "channel"
                ],
                "summary": "List channel capabilities and schemas",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.ChannelMeta"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/channels/{platform}": {
            "get": {
                "description": "Get channel meta information including capabilities and schemas",
                "tags": [
                    "channel"
                ],
                "summary": "Get channel capabilities and schemas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel platform",
                        "name": "platform",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ChannelMeta"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users/me/channels/{platform}": {
            "get": {
                "description": "Get channel binding configuration for current user",
                "tags": [
                    "channel"
                ],
                "summary": "Get channel user config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel platform",
                        "name": "platform",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/channel.ChannelIdentityBinding"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update channel binding configuration for current user",
                "tags": [
                    "channel"
                ],
                "summary": "Update channel user config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel platform",
                        "name": "platform",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Channel user config payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/channel.UpsertChannelIdentityConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/channel.ChannelIdentityBinding"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "channel.ChannelCapabilities": {
            "type": "object",
            "properties": {
                "attachments": {
                    "type": "boolean"
                },
                "block_streaming": {
                    "type": "boolean"
                },
                "buttons": {
                    "type": "boolean"
                },
                "chat_types": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "edit": {
                    "type": "boolean"
                },
                "markdown": {
                    "type": "boolean"
                },
                "media": {
                    "type": "boolean"
                },
                "native_commands": {
                    "type": "boolean"
                },
                "polls": {
                    "type": "boolean"
                },
                "reactions": {
                    "type": "boolean"
                },
                "reply": {
                    "type": "boolean"
                },
                "rich_text": {
                    "type": "boolean"
                },
                "streaming": {
                    "type": "boolean"
                },
                "text": {
                    "type": "boolean"
                },
                "threads": {
                    "type": "boolean"
                },
                "unsend": {
                    "type": "boolean"
                }
            }
        },
        "channel.ChannelConfig": {
            "type": "object",
            "properties": {
                "botID": {
                    "type": "string"
                },
                "channelType": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "credentials": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "externalIdentity": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "routing": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "selfIdentity": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "verifiedAt": {
                    "type": "string"
                }
            }
        },
        "channel.ChannelIdentityBinding": {
            "type": "object",
            "properties": {
                "channelIdentityID": {
                    "type": "string"
                },
                "channelType": {
                    "type": "string"
                },
                "config": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "channel.ConfigSchema": {
            "type": "object",
            "properties": {
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/channel.FieldSchema"
                    }
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "channel.FieldSchema": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "enum": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "example": {},
                "required": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/channel.FieldType"
                }
            }
        },
        "channel.FieldType": {
            "type": "string",
            "enum": [
                "string",
                "secret",
                "bool",
                "number",
                "enum"
            ],
            "x-enum-varnames": [
                "FieldString",
                "FieldSecret",
                "FieldBool",
                "FieldNumber",
                "FieldEnum"
            ]
        },
...
        "channel.TargetHint": {
            "type": "object",
            "properties": {
                "example": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                }
            }
        },
        "channel.TargetSpec": {
            "type": "object",
            "properties": {
                "format": {
                    "type": "string"
                },
                "hints": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/channel.TargetHint"
                    }
                }
            }
        },
...
        "channel.UpsertChannelIdentityConfigRequest": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "object",
                    "additionalProperties": {}
                }
            }
        },
        "channel.UpsertConfigRequest": {
            "type": "object",
            "properties": {
                "credentials": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "external_identity": {
                    "type": "string"
                },
                "routing": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "self_identity": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "status": {
                    "type": "string"
                },
                "verified_at": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ChannelMeta": {
            "type": "object",
            "properties": {
                "capabilities": {
                    "$ref": "#/definitions/channel.ChannelCapabilities"
                },
                "config_schema": {
                    "$ref": "#/definitions/channel.ConfigSchema"
                },
                "configless": {
                    "type": "boolean"
                },
                "display_name": {
                    "type": "string"
                },
                "target_spec": {
                    "$ref": "#/definitions/channel.TargetSpec"
                },
                "type": {
                    "type": "string"
                },
                "user_config_schema": {
                    "$ref": "#/definitions/channel.ConfigSchema"
                }
            }
        },
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "handlers.LoginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "token_type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}